# swagger.yaml
openapi: 3.0.0
info:
  title: Payment Service API
  description: |
    Payment processing service for ShopCx demo application.
  version: 1.0.0
  contact:
    name: Security Team
    email: security@example.com

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /api/payments/process:
    post:
      summary: Process a payment
      description: Process a payment transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  description: Payment amount
                cardNumber:
                  type: string
                  description: Credit card number
                cvv:
                  type: string
                  description: Card verification value
                additionalData:
                  type: object
                  description: Additional data
      responses:
        '200':
          description: Payment processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                    description: JWT token containing payment details
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/payments/verify/{token}:
    get:
      summary: Verify a payment token
      description: Verify a payment token.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: JWT token to verify
      responses:
        '200':
          description: Token verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  amount:
                    type: number
                  cardNumber:
                    type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/payments/refund:
    post:
      summary: Process a refund
      description: Process a refund for a transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                transactionId:
                  type: string
                  description: ID of the transaction to refund
                amount:
                  type: number
                  description: Refund amount
                settings:
                  type: object
                  description: Additional settings
      responses:
        '200':
          description: Refund processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/admin/transactions/clear:
    post:
      summary: Clear all transactions
      description: Clear all transaction records.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                profile:
                  type: object
                  description: Admin profile
      responses:
        '200':
          description: All transactions cleared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Detailed error message

    Success:
      type: object
      properties:
        status:
          type: string
          enum: [success]
          description: Success status 